shems/requirements.txt
aniso8601==9.0.1
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
PyMySQL==1.1.0
python-ulid==1.1.0
pytz==2023.3.post1
six==1.16.0
Werkzeug==3.0.1
zipp==3.17.0
